# Source/Sheekr.WebApi/wwwroot/swagger/v1/swagger.json
{
  "x-generator": "NSwag v12.0.14.0 (NJsonSchema v9.13.18.0 (Newtonsoft.Json v11.0.0.0))",
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "consumes": [
    "application/json-patch+json",
    "application/json",
    "text/json",
    "application/*+json"
  ],
  "paths": {
    "/api/Alunos": {
      "get": {
        "tags": [
          "Alunos"
        ],
        "operationId": "Alunos_GetAll",
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Alunos"
        ],
        "operationId": "Alunos_Post",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "command",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CriarAlunoCommand"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/api/Alunos/{id}": {
      "get": {
        "tags": [
          "Alunos"
        ],
        "operationId": "Alunos_Get",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Alunos"
        ],
        "operationId": "Alunos_Put",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "command",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AtualizarAlunoCommand"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Alunos"
        ],
        "operationId": "Alunos_Delete",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/api/Designacoes": {
      "get": {
        "tags": [
          "Designacoes"
        ],
        "operationId": "Designacoes_GetAll",
        "parameters": [
          {
            "type": "integer",
            "name": "qtde",
            "in": "query",
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Designacoes"
        ],
        "operationId": "Designacoes_Post",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "command",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DesignarCommand"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/api/Designacoes/GerarPdf": {
      "get": {
        "tags": [
          "Designacoes"
        ],
        "operationId": "Designacoes_GerarPdf",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "command",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PdfGenerateCommand"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/api/Designacoes/{id}": {
      "get": {
        "tags": [
          "Designacoes"
        ],
        "operationId": "Designacoes_Get",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Designacoes"
        ],
        "operationId": "Designacoes_Put",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "command",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AtualizarDesignacaoCommand"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Designacoes"
        ],
        "operationId": "Designacoes_Delete",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/api/Licoes": {
      "get": {
        "tags": [
          "Licoes"
        ],
        "operationId": "Licoes_GetAll",
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Licoes"
        ],
        "operationId": "Licoes_Post",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "command",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CriarLicaoCommand"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/api/Licoes/{id}": {
      "get": {
        "tags": [
          "Licoes"
        ],
        "operationId": "Licoes_Get",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Licoes"
        ],
        "operationId": "Licoes_Put",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "command",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AtualizarLicaoCommand"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Licoes"
        ],
        "operationId": "Licoes_Delete",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/api/Publicadores": {
      "get": {
        "tags": [
          "Publicadores"
        ],
        "operationId": "Publicadores_GetAll",
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Publicadores"
        ],
        "operationId": "Publicadores_Post",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "command",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CriarPublicadorCommand"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/api/Publicadores/{id}": {
      "get": {
        "tags": [
          "Publicadores"
        ],
        "operationId": "Publicadores_Get",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Publicadores"
        ],
        "operationId": "Publicadores_Put",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "command",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AtualizarPublicadorCommand"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Publicadores"
        ],
        "operationId": "Publicadores_Delete",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "CriarAlunoCommand": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "alunoId",
        "publicadorId",
        "fazLeitura",
        "fazDemonstracao",
        "fazDiscurso"
      ],
      "properties": {
        "alunoId": {
          "type": "integer",
          "format": "int32"
        },
        "publicadorId": {
          "type": "integer",
          "format": "int32"
        },
        "fazLeitura": {
          "type": "boolean"
        },
        "fazDemonstracao": {
          "type": "boolean"
        },
        "fazDiscurso": {
          "type": "boolean"
        }
      }
    },
    "AtualizarAlunoCommand": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "alunoId",
        "publicadorId",
        "fazLeitura",
        "fazDemonstracao",
        "fazDiscurso"
      ],
      "properties": {
        "alunoId": {
          "type": "integer",
          "format": "int32"
        },
        "publicadorId": {
          "type": "integer",
          "format": "int32"
        },
        "fazLeitura": {
          "type": "boolean"
        },
        "fazDemonstracao": {
          "type": "boolean"
        },
        "fazDiscurso": {
          "type": "boolean"
        }
      }
    },
    "PdfGenerateCommand": {
      "type": "object",
      "additionalProperties": false,
      "properties": {
        "nomeArquivo": {
          "type": "string"
        },
        "designacoes": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    },
    "DesignarCommand": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "designacaoId",
        "licaoId",
        "alunoPrincipalId",
        "data",
        "tipo",
        "local"
      ],
      "properties": {
        "designacaoId": {
          "type": "integer",
          "format": "int32"
        },
        "licaoId": {
          "type": "integer",
          "format": "int32"
        },
        "alunoPrincipalId": {
          "type": "integer",
          "format": "int32"
        },
        "alunoAjudanteId": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "string",
          "format": "date-time"
        },
        "tipo": {
          "$ref": "#/definitions/TipoDesignacaoEscola"
        },
        "local": {
          "$ref": "#/definitions/LocalDesignacao"
        }
      }
    },
    "TipoDesignacaoEscola": {
      "type": "integer",
      "description": "",
      "x-enumNames": [
        "Leitura",
        "PrimeiraConversa",
        "PrimeiraRevisita",
        "SegundaRevisita",
        "EstudoBiblico",
        "Discurso"
      ],
      "enum": [
        1,
        2,
        3,
        4,
        5,
        6
      ]
    },
    "LocalDesignacao": {
      "type": "integer",
      "description": "",
      "x-enumNames": [
        "SalaoPrincipal",
        "SalaB",
        "SalaC"
      ],
      "enum": [
        1,
        2,
        3
      ]
    },
    "AtualizarDesignacaoCommand": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "designacaoId",
        "licaoId",
        "alunoPrincipalId",
        "data",
        "tipo",
        "local"
      ],
      "properties": {
        "designacaoId": {
          "type": "integer",
          "format": "int32"
        },
        "licaoId": {
          "type": "integer",
          "format": "int32"
        },
        "alunoPrincipalId": {
          "type": "integer",
          "format": "int32"
        },
        "alunoAjudanteId": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "string",
          "format": "date-time"
        },
        "tipo": {
          "$ref": "#/definitions/TipoDesignacaoEscola"
        },
        "local": {
          "$ref": "#/definitions/LocalDesignacao"
        }
      }
    },
    "CriarLicaoCommand": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "titulo": {
          "type": "string"
        }
      }
    },
    "AtualizarLicaoCommand": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "titulo": {
          "type": "string"
        }
      }
    },
    "CriarPublicadorCommand": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "publicadorId",
        "sexo",
        "privilegio"
      ],
      "properties": {
        "publicadorId": {
          "type": "integer",
          "format": "int32"
        },
        "primeiroNome": {
          "type": "string"
        },
        "ultimoNome": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "telefone": {
          "type": "string"
        },
        "sexo": {
          "$ref": "#/definitions/Genero"
        },
        "privilegio": {
          "$ref": "#/definitions/Privilegio"
        }
      }
    },
    "Genero": {
      "type": "integer",
      "description": "",
      "x-enumNames": [
        "Masculino",
        "Feminino"
      ],
      "enum": [
        1,
        2
      ]
    },
    "Privilegio": {
      "type": "integer",
      "description": "",
      "x-enumNames": [
        "Nenhum",
        "ServoMinisterial",
        "Anciao"
      ],
      "enum": [
        1,
        2,
        3
      ]
    },
    "AtualizarPublicadorCommand": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "publicadorId",
        "sexo",
        "privilegio"
      ],
      "properties": {
        "publicadorId": {
          "type": "integer",
          "format": "int32"
        },
        "primeiroNome": {
          "type": "string"
        },
        "ultimoNome": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "telefone": {
          "type": "string"
        },
        "sexo": {
          "$ref": "#/definitions/Genero"
        },
        "privilegio": {
          "$ref": "#/definitions/Privilegio"
        }
      }
    }
  }
}